Fixed incorrect handling of error responses
According to the WAA->WLS communication protocol responses to requests
may be of any protocol version not exceeding the one specified in the
request. This means that, particularly with error responses, an older
response type may be returned which does not contain all the parameters
expected in the response. Previously, this caused erroneous errors to
be thrown in certain situations (such as the user canceling the
authorisation process). This commit addresses this issue.